requirements.txt
pandas
matplotlib
python-dotenv
pyyaml
mkdocs
mkdocs-material
mkdocs-autorefs 
mkdocstrings 
mkdocstrings-python
pymdown-extensions
seaborn
yfinance
plotly
kaleido
boto3
pyarrow
fastparquet
pandas_datareader
setuptools
pyyaml